+++ Libplanet.Tests/Tx/TransactionTest.cs
@@ -97,28 +97,28 @@ namespace Libplanet.Tests.Tx
         public void CreateWithDefaultUpdatedAddresses()
         {
             Transaction<DumbAction> emptyTx = Transaction<DumbAction>.Create(
-                _fx.PrivateKey,
+                _fx.PrivateKey1,
                 new DumbAction[0]
             );
             Assert.Empty(emptyTx.UpdatedAddresses);
 
             var tx = Transaction<PolymorphicAction<BaseAction>>.Create(
-                _fx.PrivateKey,
+                _fx.PrivateKey1,
                 _fx.TxWithActions.Actions
             );
             Assert.Equal(
-                new[] { _fx.Address }.ToImmutableHashSet(),
+                new[] { _fx.Address1 }.ToImmutableHashSet(),
                 tx.UpdatedAddresses
             );
 
             Address additionalAddr = new PrivateKey().PublicKey.ToAddress();
             var txWithAddr = Transaction<PolymorphicAction<BaseAction>>.Create(
-                _fx.PrivateKey,
+                _fx.PrivateKey1,
                 _fx.TxWithActions.Actions,
                 new[] { additionalAddr }.ToImmutableHashSet()
             );
             Assert.Equal(
-                new[] { _fx.Address, additionalAddr }.ToHashSet(),
+                new[] { _fx.Address1, additionalAddr }.ToHashSet(),
                 txWithAddr.UpdatedAddresses.ToHashSet()
             );
         }
@@ -128,7 +128,7 @@ namespace Libplanet.Tests.Tx
         {
             DateTimeOffset rightBefore = DateTimeOffset.UtcNow;
             Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
-                _fx.PrivateKey,
+                _fx.PrivateKey1,
                 new DumbAction[0],
                 ImmutableHashSet<Address>.Empty
             );
@@ -153,7 +153,7 @@ namespace Libplanet.Tests.Tx
             // The actions parameter cannot be null.
             Assert.Throws<ArgumentNullException>(() =>
                 Transaction<DumbAction>.Create(
-                    _fx.PrivateKey,
+                    _fx.PrivateKey1,
                     null,
                     ImmutableHashSet<Address>.Empty,
                     DateTimeOffset.UtcNow
@@ -167,7 +167,7 @@ namespace Libplanet.Tests.Tx
             var action = new ThrowException { Throw = true };
             Assert.Throws<UnexpectedlyTerminatedTxRehearsalException>(() =>
                 Transaction<ThrowException>.Create(
-                    _fx.PrivateKey,
+                    _fx.PrivateKey1,
                     new[] { action },
                     ImmutableHashSet<Address>.Empty,
                     DateTimeOffset.UtcNow
@@ -602,7 +602,7 @@ namespace Libplanet.Tests.Tx
                 new DumbAction(addresses[2], "R", true, recordRandom: true),
             };
             Transaction<DumbAction> tx =
-                Transaction<DumbAction>.Create(_fx.PrivateKey, actions);
+                Transaction<DumbAction>.Create(_fx.PrivateKey1, actions);
             foreach (bool rehearsal in new[] { false, true })
             {
                 DumbAction.RehearsalRecords.Value =
@@ -628,7 +628,7 @@ namespace Libplanet.Tests.Tx
                 {
                     ActionEvaluation<DumbAction> eval = evaluations[i];
                     Assert.Equal(actions[i], eval.Action);
-                    Assert.Equal(_fx.Address, eval.InputContext.Signer);
+                    Assert.Equal(_fx.Address1, eval.InputContext.Signer);
                     Assert.Equal(addresses[0], eval.InputContext.Miner);
                     Assert.Equal(1, eval.InputContext.BlockIndex);
                     Assert.Equal(rehearsal, eval.InputContext.Rehearsal);
@@ -843,12 +843,12 @@ namespace Libplanet.Tests.Tx
         {
             var actions = new List<DumbAction>();
             Transaction<DumbAction> t1 = Transaction<DumbAction>.Create(
-                _fx.PrivateKey,
+                _fx.PrivateKey1,
                 actions
             );
             var t2 = new Transaction<DumbAction>(
-                _fx.PrivateKey.PublicKey.ToAddress(),
-                _fx.PrivateKey.PublicKey,
+                _fx.PrivateKey1.PublicKey.ToAddress(),
+                _fx.PrivateKey1.PublicKey,
                 ImmutableHashSet<Address>.Empty,
                 t1.Timestamp,
                 actions,